Delete unusable components (#54)
