+++ client/src/App.js
@@ -1,7 +1,7 @@
 import {Switch, Route} from 'react-router-dom'
 import React from 'react'
 import './App.css'
-import SignIn from './components/SignIn/sign-in'
+import SignIn from './components/SignIn/SignIn'
 import ProtectedRoute from './components/ProtectedRoute/ProtectedRoute'
 import Register from './components/Register/Register'
