--- client/src/components/Header/Header.css
@@ -1,8 +1,0 @@
-.tab-container {
-  display: flex;
-  list-style: none;
-}
-
-.tabs {
-  margin-right: 20px;
-}
