+++ client/src/components/Profile/Profile.js
@@ -3,7 +3,7 @@ import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
 import ProfileCard from './ProfileCard'
 import { makeStyles } from '@material-ui/core/styles'
 import Grid from '@material-ui/core/Grid'
-import TopMenu from '../TopMenu/top-menu'
+import TopMenu from '../TopMenu/TopMenu'
 import ProfileTabs from './ProfileTabs'
 
 const useStyles = makeStyles(theme => ({
