+++ client/src/components/Wall/Wall.js
@@ -3,7 +3,7 @@ import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
 import { makeStyles } from '@material-ui/core/styles'
 import CircularProgress from '@material-ui/core/CircularProgress'
 import Grid from '@material-ui/core/Grid'
-import TopMenu from '../TopMenu/top-menu'
+import TopMenu from '../TopMenu/TopMenu'
 import {useDispatch, useSelector} from 'react-redux'
 import Tape from '../Tape/Tape'
 import {fetchWallPosts} from '../../actions/postActions'
